+++ extension/CHANGELOG.md
@@ -22,6 +22,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     - `NAB: Import Translations from .csv` imports and updates targets of selected XLF file from a .csv file. View [README](README.md) for full documentation.
     - Related settings:
       - `NAB.XliffCSVExportPath` sets the export path for `NAB: Export Translations to .csv`. Default path for export is the Translation file directory.
+- Fixed issues
+  - Fixed issue #130, "NAB: Find code source of current line" broken if there's special characters in name
+
 ## [1.0.0] Out of preview - 2021-03-23
 
 We're out of preview no more beta!